+++ CHANGELOG.md
@@ -17,7 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-*
+* Support no `faces` in a dimension from spec in ´getDimension`
 
 ## [2.25.0] - 2022-03-24